backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM Integration
openai==1.10.0
anthropic==0.8.1
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
pinecone-client==3.0.0

# Embeddings & RAG
sentence-transformers==2.3.1
rank-bm25==0.2.2

# Task Queue & Async
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
orjson==3.9.12

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Browser Automation
playwright==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
